Revisit the comments of the cluster, separate the backward compatability tests
