+++ src/app/clusters/actions-server/tests/BUILD.gn
@@ -37,16 +37,37 @@ source_set("actions-cluster-test-srcs") {
 chip_test_suite("tests") {
   output_name = "libTestCodeDrivenActionsCluster"
 
-  test_sources = [
-    "TestActionsCluster.cpp",
-    "TestCodeDrivenActionsCluster.cpp",
-  ]
+  test_sources = [ "TestActionsCluster.cpp" ]
 
   public_deps = [
-    ":actions-cluster-test-srcs",
+    "${chip_root}/src/app/clusters/actions-server",
     "${chip_root}/src/app/server-cluster/testing:testing",
-    "${chip_root}/src/app/tests:app-test-stubs",
   ]
 
   cflags = [ "-Wconversion" ]
+}
+
+chip_test_suite("tests-backwards-compatibility") {
+  import(
+      "${chip_root}/src/app/clusters/actions-server/app_config_dependent_sources.gni")
+
+  output_name = "libTestActionsClusterBackwardsCompatibility"
+
+  test_sources = [ "TestActionsClusterBackwardCompatability.cpp" ]
+
+  sources = []
+  foreach(file, app_config_dependent_sources) {
+    sources += [ "../" + file ]
+  }
+
+  sources += [ "${chip_root}/zzz_generated/app-common/app-common/zap-generated/attributes/Accessors.cpp" ]
+
+  public_deps = [
+    "${chip_root}/src/app/clusters/actions-server",
+    "${chip_root}/src/app/common:cluster-objects",
+    "${chip_root}/src/app/server-cluster/testing:testing",
+    "${chip_root}/src/app/tests:app-test-stubs",
+    "${chip_root}/src/app/util/mock:mock_codegen_data_model",
+    "${chip_root}/src/lib/core",
+  ]
 }
